+++ notify-impl/src/main/java/com/dwarfeng/notify/impl/handler/router/GroovyRouterRegistry.java
@@ -9,7 +9,6 @@ import com.dwarfeng.notify.stack.handler.Router;
 import groovy.lang.GroovyClassLoader;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.config.ConfigurableBeanFactory;
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.annotation.Scope;
@@ -33,11 +32,11 @@ public class GroovyRouterRegistry extends AbstractRouterRegistry {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(GroovyRouterRegistry.class);
 
-    @Autowired
-    private ApplicationContext ctx;
+    private final ApplicationContext ctx;
 
-    public GroovyRouterRegistry() {
+    public GroovyRouterRegistry(ApplicationContext ctx) {
         super(ROUTER_TYPE);
+        this.ctx = ctx;
     }
 
     @Override